Разбивает код на модули, уточняет константы
